ScrollingFrame
This class was made for a functioning scrollbar (Frame) in python, while using grid. The different functions lets the program react to events such as entering and leaving the frame, using the mouse-wheel for scrolling and resizing the scrollbar after resizing the program.
An example code for using this class is also provided.